reorganise top level files
